% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERiskProfileR.R
\name{as_df}
\alias{as_df}
\title{Convert Pregnancy List to Data Table}
\usage{
as_df(res)
}
\arguments{
\item{res}{A list of \code{Pregnancy} R6 objects.}
}
\value{
A \code{data.table} containing the aggregated data from all objects,
with missing columns filled where necessary.
}
\description{
An internal helper function that takes a list of \code{Pregnancy} objects and
collapses them into a single \code{data.table}.
}
\keyword{internal}
