% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_risk_calculation.R
\name{extract_risk_scores_from_file}
\alias{extract_risk_scores_from_file}
\title{Extract Risk Scores from HTML File}
\usage{
extract_risk_scores_from_file(html_file, report_as_text = FALSE)
}
\arguments{
\item{html_file}{String, path to HTML file}

\item{report_as_text}{Boolean, whether to return risks as text fractions (default: FALSE)}
}
\value{
List containing risk scores and MoM values (see extract_risk_scores)
}
\description{
Reads an HTML file and extracts preeclampsia risk scores and biomarker MoM values
}
\seealso{
\code{\link{extract_risk_scores}}
}
\keyword{online}
