% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_risk_calculation.R
\name{get_expected_map}
\alias{get_expected_map}
\title{Calculate Expected Mean Arterial Pressure}
\usage{
get_expected_map(form_data, truncate = TRUE)
}
\arguments{
\item{form_data}{A list containing:
\itemize{
\item ga: Numeric, gestational age in weeks
\item weight: Numeric, maternal weight in kg
\item height: Numeric, maternal height in cm
\item race: Integer (2=Afro-Caribbean)
\item smoking: Boolean, smoking status
\item chronic_hypertension: Boolean, presence of chronic hypertension
\item diabetes_type_i: Boolean, presence of Type 1 diabetes
\item diabetes_type_ii: Boolean, presence of Type 2 diabetes
\item mother_pe: Boolean, maternal history of preeclampsia
\item previous: Boolean, previous pregnancy status
\item previous_pe: Boolean, previous preeclampsia status
\item previous_interval: Numeric, interval since previous pregnancy in years
}}
}
\value{
Numeric value representing the expected MAP
}
\description{
Calculates the expected Mean Arterial Pressure (MAP) based on maternal
characteristics and medical history
}
\keyword{local}
