% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_risk_calculation.R
\name{get_expected_utpi}
\alias{get_expected_utpi}
\title{Calculate Expected Uterine Artery Pulsatility Index}
\usage{
get_expected_utpi(form_data, truncate = TRUE)
}
\arguments{
\item{form_data}{A list containing maternal characteristics including:
\itemize{
\item ga: Numeric, gestational age in weeks
\item weight: Numeric, maternal weight in kg
\item age: Numeric, maternal age in years
\item race: Integer (2=Afro-Caribbean, 4=East Asian, 5=Mixed)
\item diabetes_type_i: Boolean, presence of Type 1 diabetes
\item previous: Boolean, previous pregnancy status
\item previous_pe: Boolean, previous preeclampsia status
}}
}
\value{
Numeric value representing the expected UtAPI
}
\description{
Calculates the expected Uterine Artery Pulsatility Index (UtAPI) based on
maternal characteristics
}
\keyword{local}
