% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_risk_calculation.R
\name{get_mom_map}
\alias{get_mom_map}
\title{Calculate MAP MoM}
\usage{
get_mom_map(form_data)
}
\arguments{
\item{form_data}{A list containing:
\itemize{
\item map: Numeric, measured MAP value
}}
}
\value{
Numeric value representing MAP MoM
}
\description{
Calculates the Multiple of Median (MoM) for Mean Arterial Pressure
}
\keyword{local}
