% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_risk_calculation.R
\name{get_mom_utpi}
\alias{get_mom_utpi}
\title{Calculate UtAPI MoM}
\usage{
get_mom_utpi(form_data, truncation = TRUE)
}
\arguments{
\item{form_data}{A list containing:
\itemize{
\item utpi: Numeric, measured UtAPI value
}}
}
\value{
Numeric value representing UtAPI MoM
}
\description{
Calculates the Multiple of Median (MoM) for Uterine Artery Pulsatility Index
}
\keyword{local}
