% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_risk_calculation.R
\name{get_p_MAP_PI}
\alias{get_p_MAP_PI}
\title{Calculate MAP-PI Joint Distribution}
\usage{
get_p_MAP_PI(mom_MAP, mom_PI, g)
}
\arguments{
\item{mom_MAP}{Numeric, MAP MoM value}

\item{mom_PI}{Numeric, PI MoM value}

\item{g}{Numeric vector, gestational ages}
}
\value{
Numeric vector of probability densities
}
\description{
Calculates the joint probability distribution for MAP and PI MoM values
}
