% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_risk_calculation.R
\name{get_p_MAP_PI_PlGF}
\alias{get_p_MAP_PI_PlGF}
\title{Calculate MAP-PI-PlGF Joint Distribution}
\usage{
get_p_MAP_PI_PlGF(mom_MAP, mom_PI, mom_PlGF, g)
}
\arguments{
\item{mom_MAP}{Numeric, MAP MoM value}

\item{mom_PI}{Numeric, PI MoM value}

\item{mom_PlGF}{Numeric, PlGF MoM value}

\item{g}{Numeric vector, gestational ages}
}
\value{
Numeric vector of probability densities
}
\description{
Calculates the joint probability distribution for MAP, PI, and PlGF MoM values
}
