% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_risk_calculation.R
\name{get_p_prior}
\alias{get_p_prior}
\title{Calculate Prior Probability}
\usage{
get_p_prior(g, form_data, mom_MAP = NA, mom_PI = NA, mom_PlGF = NA)
}
\arguments{
\item{g}{Numeric vector, gestational ages}

\item{form_data}{List containing maternal characteristics}

\item{mom_MAP}{Numeric, MAP MoM value (default: NA)}

\item{mom_PI}{Numeric, PI MoM value (default: NA)}

\item{mom_PlGF}{Numeric, PlGF MoM value (default: NA)}
}
\value{
Numeric vector of probability densities
}
\description{
Calculates prior probability combining maternal characteristics and biomarker information
}
