% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERiskProfileR-package.R
\docType{package}
\name{PERiskProfileR-package}
\alias{PERiskProfileR}
\alias{PERiskProfileR-package}
\title{PERiskProfileR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andjar.github.io/PERiskProfileR/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Other contributors:
\itemize{
  \item  [contributor]
}

}
\keyword{internal}
