% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{PERiskProfileRStartupMessage}
\alias{PERiskProfileRStartupMessage}
\title{Generate Package Startup Message}
\usage{
PERiskProfileRStartupMessage()
}
\description{
Generate Package Startup Message
}
\keyword{internal}
