% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{get_demo_data}
\alias{get_demo_data}
\title{Retrieve Demo Dataset}
\usage{
get_demo_data()
}
\value{
A \code{data.table} object containing the contents of the demo dataset.
}
\description{
This function provides access to a demo dataset included in the package.
The dataset is stored as a CSV file in the \code{extdata} directory of the package.
}
\examples{
\dontrun{
# Load the demo dataset
demo_data <- get_demo_data()
head(demo_data)
}

}
\keyword{prepare}
