% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{get_validation_data}
\alias{get_validation_data}
\title{Retrieve Validation Dataset}
\usage{
get_validation_data()
}
\value{
A \code{data.table} object containing the contents of the validation dataset.
}
\description{
This function provides access to a validation dataset included in the package.
The dataset is stored as a CSV file in the \code{extdata} directory of the package.
}
\keyword{prepare}
