% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERiskProfileR-package.R
\docType{package}
\name{PERiskProfileR-package}
\alias{PERiskProfileR}
\alias{PERiskProfileR-package}
\title{PERiskProfileR: Calculate Preeclampsia Risk Scores in First Trimester}
\description{
This package is based on the published Fetal Medicine Foundation (FMF) Risk Calculator for preeclampsia, based on first trimester parameters. Risk scores can be calculated either locally or by automatically submitting data to the web form.
}
\author{
\strong{Maintainer}: Anders Hagen Jarmund \email{anders.h.jarmund@ntnu.no} (\href{https://orcid.org/0000-0002-3923-1637}{ORCID})

}
\keyword{internal}
